// src/app/api/models/invoice-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InvoiceLineRequest } from '../models/invoice-line-request';

/**
 * A held sale is written through this. Completing and voiding are actions.
 */
export interface InvoiceRequest {
  customer: number;
  hold_label?: string;
  lines?: Array<InvoiceLineRequest>;
  quotation?: number | null;
  walk_in_name?: string;
  walk_in_phone?: string;
}