// src/app/api/models/tender-input-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CurrencyEnum } from '../models/currency-enum';
import { TenderKindEnum } from '../models/tender-kind-enum';
export interface TenderInputRequest {
  amount: string;
  currency?: CurrencyEnum;
  kind: TenderKindEnum;
  reference?: string;
}